requirements.txt
undetected-chromedriver==3.5.4
playwright==1.40.0
playwright-stealth==1.0.6
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
cryptography==41.0.8
boto3==1.34.0
pyotp==2.9.0
pyyaml==6.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
fake-useragent==1.4.0
fingerprint-suite==1.2.0
schedule==1.2.0
loguru==0.7.2
rich==13.7.0
typer==0.9.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
celery==5.3.4
